Move physics components and Operator
Moved into a physics sub-package of component.  Without breaking up the
component package like this it's going to balloon out of control pretty
quickly.
