+++ core/src/com/brinkman/platformer/component/physics/PhysicsComponent.java
@@ -1,5 +1,6 @@
-package com.brinkman.platformer.component;
+package com.brinkman.platformer.component.physics;
 
+import com.brinkman.platformer.component.RootComponent;
 import com.brinkman.platformer.physics.Body;
 
 /**
