+++ core/src/com/brinkman/platformer/entity/actor/Exit.java
@@ -2,8 +2,8 @@ package com.brinkman.platformer.entity.actor;
 
 import com.badlogic.gdx.utils.Logger;
 import com.brinkman.platformer.GameWorld;
-import com.brinkman.platformer.component.ControlledPhysicsComponent;
-import com.brinkman.platformer.component.PhysicsComponent;
+import com.brinkman.platformer.component.physics.ControlledPhysicsComponent;
+import com.brinkman.platformer.component.physics.PhysicsComponent;
 import com.brinkman.platformer.component.RootComponent;
 import com.brinkman.platformer.entity.Entity;
 import com.brinkman.platformer.entity.actor.item.Item;
