+++ core/src/com/brinkman/platformer/entity/actor/platform/Platform.java
@@ -2,6 +2,8 @@ package com.brinkman.platformer.entity.actor.platform;
 
 import com.badlogic.gdx.graphics.g2d.TextureRegion;
 import com.brinkman.platformer.component.*;
+import com.brinkman.platformer.component.physics.ControlledPhysicsComponent;
+import com.brinkman.platformer.component.physics.PhysicsComponent;
 import com.brinkman.platformer.entity.actor.Actor;
 import com.brinkman.platformer.entity.actor.Player;
 import com.brinkman.platformer.util.Constants;
