+++ core/src/com/brinkman/platformer/input/ControllerProcessor.java
@@ -6,8 +6,7 @@ import com.badlogic.gdx.controllers.Controllers;
 import com.badlogic.gdx.controllers.PovDirection;
 import com.badlogic.gdx.math.Vector3;
 import com.badlogic.gdx.utils.Logger;
-import com.brinkman.platformer.component.ControlledPhysicsComponent;
-import com.brinkman.platformer.component.PhysicsComponent;
+import com.brinkman.platformer.component.physics.PhysicsComponent;
 import com.brinkman.platformer.entity.actor.Player;
 import com.brinkman.platformer.physics.ControlledBody;
 import com.brinkman.platformer.util.ControllerMappings;
